trip service

+++ identity-service/src/main/java/com/example/demo/dto/response/ApiResponse.java
@@ -12,7 +12,7 @@ import lombok.experimental.FieldDefaults;
 public class ApiResponse<T> {
     @Builder.Default
     int code = 1000;
-    int count;
+    Long count;
     String message;
     T result;
 }
