+++ locationservice/src/main/java/com/rideup/locationservice/dto/response/ApiResponse.java
@@ -13,7 +13,7 @@ import lombok.experimental.FieldDefaults;
 public class ApiResponse<T> {
     @Builder.Default
     int code = 1000;
-    int count;
+    Long count;
     String message;
     T result;
 }
